requirements.txt
pandas
transformers
streamlit
matplotlib
torch
sentencepiece
plotly
scipy
nltk
scikit-learn
streamlit-toggle-switch
xgboost==1.5.0

